Check for user before rendering circuit sapling
The circuit sapling will now check if the user is logged in before
rendering. If the user is not logged in, it will redirect them to
the login page.

This fixes an issue where if the user navigates to the circuits page
directly they would occasionally be taken to the page without being
asked to log in and the page would not render the Canopy navbar. This
was caused by a race condition where both the login page and the
circuits page would render simultaneously, causing the page that
finished rendering last to overwrite the first one. This was not an
issue if the login paged rendered last. However, if the circuits page
rendered last the page would render incorrectly. This commit
fixes that issue by forcing the circuits page to redirect the user
in the case that it is the page that finishes rendering last.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>