Fix key table action options
Modify action button in the key table so that the expanding action
options move with the action column.

Also modify css to allow the expanded actions options of the last key
in the table to go outside of the table boundary.

Previously, if the browser window was smaller than the table width,
requiring horizontal scrolling, when hovering the cursor over the action
button the expanded action options were off screen.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>