Update circuit sapling styles
This updates the circuit sapling styles so that the sapling displays
correctly in Grid. These changes consist of updates to certain colors to
ensure proper display and contrast on both white backgrounds (splinter
ui) and on light grey backgrounds (grid). This also fixes a couple small
issues with spacing and table layout This does not update the sapling to
match the Grid design system. That work will need further discussions.

Signed-off-by: Davey Newhall <newhall@bitwise.io>