Add permission services
Add an `alpha-node-permissions` and a `beta-node-permissions` service
to the oauth docker-compose file. These services use the "splinter user
list" and "splinter authid create" cli commands to assign the admin role
to any user that logs into the node.

Additionally, lower the verbosity of splinterd to avoid having the logs
clogged up by the actix debug logs that print every time a user is
authenticated. These services executes the "splinter user list"
command every 5 seconds which causes a log to be printed reporting that
the user was authenticated.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>