Update biome profile info
Previously when the biome example for the splinter UI was built, the
biome/profile endpoint would return a 404 when queried because it did
not exist as an endpoint with the splinterd configuration used. Now
that the endpoint exists for the configuration but profiles are only
created for OAuth users, when the splinter UI queries the endpoint, a
500 is returned because no profile exists. This commit updates the case
in the switch statement from 404 to 500 so that the profile page will
display the user's name.

Signed-off-by: Isabel Tomb <tomb@bitwise.io>